Extract cursor and put together in one cursor.rs file
