+++ packages/yew-style-in-rs-macro/src/style/css.rs
@@ -1,66 +1,12 @@
 use proc_macro2::TokenStream;
 use quote::quote;
 use std::io::Write;
-use std::iter::Peekable;
-use std::str::Chars;
 use yew_style_in_rs_core::ast::RuntimeCss;
 use yew_style_in_rs_core::transpiler::TranspiledCss;
 
+use crate::cursor::*;
 use crate::state::*;
 use crate::style::keyframes::*;
-
-// Parser for `##anim_name##` in code.
-#[derive(Clone, Debug)]
-pub struct Cursor<'a> {
-    content: Peekable<Chars<'a>>,
-}
-impl<'a> Cursor<'a> {
-    fn new(content: &'a str) -> Self {
-        Self {
-            content: content.chars().peekable(),
-        }
-    }
-
-    fn is_empty(&mut self) -> bool {
-        self.content.peek().is_none()
-    }
-
-    fn peek(&mut self, ch: char) -> bool {
-        self.content.peek() == Some(&ch)
-    }
-
-    fn next(&mut self) -> Option<char> {
-        self.content.next()
-    }
-
-    fn take(&mut self, ch: char) -> Option<char> {
-        if let Some(c) = self.content.next() {
-            if c == ch {
-                Some(ch)
-            } else {
-                None
-            }
-        } else {
-            None
-        }
-    }
-
-    fn take_until(&mut self, delimiter: char) -> Option<String> {
-        let mut ret = String::new();
-        loop {
-            if let Some(&c) = self.content.peek() {
-                if c == delimiter {
-                    return Some(ret);
-                } else {
-                    ret.push(c);
-                    self.content.next();
-                }
-            } else {
-                return None;
-            }
-        }
-    }
-}
 
 // replace animation name to animation name with id
 fn replace_animation_name(
@@ -75,9 +21,9 @@ fn replace_animation_name(
             cursor.take('#').unwrap();
             if cursor.peek('#') {
                 cursor.take('#').unwrap();
-                let name = cursor
-                    .take_until('#')
-                    .ok_or("`##<animation_name>##` is expected")?;
+                let (name, _) = cursor
+                    .take_until(&['#'])
+                    .map_err(|_| "`##<animation_name>##` is expected")?;
 
                 if let Some(name) = animation_names.iter().find(|n| n.animation_name == name) {
                     code += &name.animation_name_with_scoped_id;
